Changed unsigned into size_t
git-svn-id: svn://svn.code.sf.net/p/yodl/svn/branches/rm_unsigned@85 8f49d26b-2354-4b37-89f9-c6476e23a9d3

+++ yodl/src/hashmap/hashmapkeys.c
@@ -2,7 +2,7 @@
 
 void hashmap_keys(register HashMap *map)
 {
-    unsigned idx;
+    size_t idx;
 
     fprintf(stderr, "There are %u active keys\n", map->d_n);
 
@@ -10,7 +10,7 @@ void hashmap_keys(register HashMap *map)
     {
         register HashItem *atIdx = map->d_map[idx];
 
-        switch ((unsigned)atIdx)
+        switch ((size_t)atIdx)
         {
             case FREE:
             continue;
